requirements.txt
SpeechRecognition
gTTS
google-generativeai
python-dotenv
pyaudio
requests